Adding support for DSCP configuration
Allow the user to do DSCP mapping for network objects
Interface and SriovPF. The feature needs the handling
of netlink messages. The pyroute2 module is used in
the encode/decode of the netlink messages.
The persistence of the DSCP configurations are taken care
by the os-net-config-dcb-config service. The user configurations
are preserved in /var/lib/os-net-config/dcb_config.yaml

The details of the DSCP configurations could be read anytime
with the cli.
``os-net-config-dcb --show``

The new DSCP configurations can be applied with cli using
``os-net-config-dcb --config <config.yaml>``

The sample config.yaml for updating DCB only is available at
etc/os-net-config/samples/dcb_config_sample.yaml