Write sriov_config.yaml for nmstate provider
In case of SR-IOV deployments using nmstate provider, the
sriov_config.yaml is not written for PFs. Add the sriov_pf
to be written, so that the derivation PCI passthrough list could
be automated. This is a temporary requirement and once the automated
scripts could use other means of deriving the PCI passthrough list,
the sriov_config.yaml shall be removed.

(cherry picked from commit d25e28d95d35419176e6824336991f9e3114dc89)