Allow disabling sriov_drivers_autoprobe
By default, VFs are automatically bound to their default kernel driver
after being created.  This leads to creating a new network interface
(linux netdev) for every VF.  Every netdev consumes IRQs (for RX/TX
queues and other operations).  This can cause a very large number of IRQ
to be created alongside which all need to be handled by a host CPU.
A single VF may consume more than 10 IRQs, depending on the driver.

Intel CPU APICs can only handle a limited amount of IRQ.  Basically,
there are 256 interrupt vectors in the APIC.  55 of them are reserved
for general purpose interrupts (exceptions, division by zero, etc.) That
leaves 201 of them for hardware/device related MSI.

Using cpu_affinity, it is possible to control which IRQs are handled by
which CPUs, up to this 201 limit.  Past that limit, the kernel will fail
to map the IRQ to a CPU if its APIC is already full.  This will cause
the IRQs to be bound to the next available CPU.

In OpenStack NFV deployments, it is frequent to use CPU isolation to
ensure that no IRQ nor CPU steal can happen on cores that are dedicated
for virtual machines.  If there are too many IRQs to pack in
housekeeping cores, there will be overspilling in CPUs that are expected
to be isolated.  This can lead to workload interruptions and in the end
cause latency and performance issues.

Additionally, even if IRQ overspilling is not an issue, all created VFs
associated with a linux network interface will be handled by
NetworkManager by default (unless some explicit configuration has been
created to force NM to ignore them).  NM will try to configure each VF
using DHCP by default.  This can cause excessive netlink churn and even
potential stalling of DPDK based applications that rely on bifurcated
drivers (e.g. mlx5).

In most cases, the VFs created by os-net-config are only meant for
virtual machines and are no use to the host.  When the VF devices are
assigned to a guest, they need to be bound to the vfio-pci driver
anyway, effectively unbinding them from their default kernel driver.
This in turn removes their associated linux netdev (and their related
IRQs).  However, until they are bound to a guest, the unused VFs consume
IRQs and can cause issues as described above.

Force sriov_drivers_autoprobe to 0 (the default value is 1) before
setting sriov_numvfs.  This causes VF PCI devices to be created but
prevents the kernel from binding them to their respective default
driver, thus preventing the creation of netdevs and avoiding the need
for CPUs to handle extraneous IRQs.

Link: https://bugzilla.redhat.com/2184735
Link: https://docs.kernel.org/PCI/pci-iov-howto.html#sr-iov-api
Link: https://github.com/nmstate/nmstate/pull/2643
Fixes: https://issues.redhat.com/browse/OSPRH-5847
Signed-off-by: Robin Jarry <rjarry@redhat.com>