Need to start DCB config service after Day 2 config is applied
When Day 2 configuration is applied the DCB config service is
not getting started. Because of this after reboot the dscp2prio
mapping is not persistent after reboot.

Changes are done to start the DCB config service when the Day 2
configuraiton is applied. With this change the issue is resolved.

Signed-off-by: Abhiram R N <abhiramrn@gmail.com>
(cherry picked from commit 5f6ba3937c808a571fed2dbdaffbea0850ccc387)