Handle nic mapping for DCB configs
As part of deployment templates the nic numbering
is being used and for DCB configs it is not supported
and expecting physical interface naming. Hence adding
changes to use the mapping of the nics to appropriate
interfaces for DCB configs, so that it can support
with nic numbering.

Signed-off-by: Abhiram R N <abhiramrn@gmail.com>
(cherry picked from commit ca52270eb56730a376a5b702619b78a088e0f8b3)