Update the VF args for OvS / OvS user bridges
The default values of the VF parameters like Trust, Spoofcheck,
Promisc are updated for OvS and OvS user bridges, Linux bonds.
This update is not effectively applied on the nmstate provider's
sriov_vf_data, leading to inconsistency between the sriov map
and the VF configurations applied. Now the update of the default
values are applied on sriov_vf_data.