sriov_config: fix race condition on boot
On boot, sriov_config.service runs in parallel with the probing of PCI
devices. There is no synchronization point that we can use to run the
service *after* all PCI devices have been probed. This leads to the
service failing depending on the time it takes for devices to be probed:

 12:31:30 os-net-config-sriov[1772]: FileNotFoundError: [Errno 2] No
  such file or directory: '/sys/class/net/ens2f0np0/device/sriov_numvfs'
 ...
 12:31:43 kernel: mlx5_core 0000:17:00.0 ens2f0np0: renamed from eth0

The issue is that there is no UDEV rule to back this up when using NIC
partitioning.

Add an explicit call to udevadm wait before trying to access the files.
If the devices have already been probed, it will not cause any delay. On
the other hand, if the devices haven't been probed yet or if the
configuration is incorrect, the command will timeout after 120 seconds
and raise an error.

Fixes: https://issues.redhat.com/browse/OSPRH-8922
Signed-off-by: Robin Jarry <rjarry@redhat.com>
(cherry picked from commit 0f2a60cb32c7c6062c581256d321a7877cc414e0)