sriov_config: fix race condition on boot
On boot, sriov_config.service runs in parallel with the probing of PCI
devices. There is no synchronisation point that we can use to run the
service *after* all PCI devices have been probed. This leads to the
service failing depending on the time it takes for devices to be probed:

 12:31:30 os-net-config-sriov[1772]: FileNotFoundError: [Errno 2] No
  such file or directory: '/sys/class/net/ens2f0np0/device/sriov_numvfs'
 ...
 12:31:43 kernel: mlx5_core 0000:17:00.0 ens2f0np0: renamed from eth0

Remove that service an only rely on UDEV rules to configure the correct
amount of VFs when the PF device appears.

Signed-off-by: Robin Jarry <rjarry@redhat.com>