sriov_config: give up on waiting for pf after 60 seconds
By default, udevadm wait has no timeout. If the -t, --timeout option is
not specified, the command will wait for the device to be initialized
forever. Here's an excerpt from udevadm(8):

 udevadm wait [options] [device|syspath] ...
     Wait for devices or device symlinks being created and initialized
     by systemd-udevd.

     -t, --timeout=SECONDS
         Maximum number of seconds to wait for the specified devices or
         device symlinks being created, initialized, or removed. The
         default value is "infinity".
                           ^^^^^^^^

If there is a misconfiguration or a hardware issue, this can lead to
stalling on boot.

Add an explicit 60 seconds timeout after which the command will fail
with an explicit error:

  Timed out for waiting devices being initialized.

Fixes: 0f2a60cb32c7 ("sriov_config: fix race condition on boot")
Reported-by: Karthik Sundaravel <ksundara@redhat.com>
Signed-off-by: Robin Jarry <rjarry@redhat.com>
(cherry picked from commit 16c9a234494ca66a0cc79994642098871636d461)