noop: Run os-net-config without applying the changes
When os-net-config is run with noop, its failing due to permission
issues or trying to write files and configure devices. When `noop`
is used, there shall be no changes made to the machine, while the
configuration files or templates derived shall be output on the
console. This is done so that the template/compatibility issues
of the config.yaml could be validated on a given hardware without
changing the system files or configuration.

(cherry picked from commit 2e60c77aee02d682028ff7795f5df1e9c9a0ae4e)
Signed-off-by: Abhiram R N <abhiramrn@gmail.com>