noop: Run os-net-config without applying the changes
When os-net-config is run with noop, its failing due to permission
issues or trying to write files and configure devices. When `noop`
is used, there shall be no changes made to the machine, while the
configuration files or templates derived shall be output on the
console. This is done so that the template/compatibility issues
of the config.yaml could be validated on a given hardware without
changing the system files or configuration.