Allow DNS changes without restarting interfaces
This change allows DNS servers and search domains to be
modified on the fly without restarting interfaces.
The DNS nameservers and search domains are written to the
ifcfg file and /etc/resolv.conf will be updated in place.

This change also includes some code cleanup and updates
string formatting in some places to match Python3 style.

Link: https://bugzilla.redhat.com/2315341
Fixes: https://issues.redhat.com/browse/OSP-32857
Fixes: https://issues.redhat.com/browse/OSP-32858
Signed-off-by: Dan Sneddon <dsneddon@redhat.com>