Fix test cases for noop
The unit testcases are failing with the error
```NameError: name 'noop' is not defined```
Invoked set_noop() in the setUp() routines of all the tests, so
that the global variable ```noop``` is defined and available to
the unit tests. os-net-config-sriov and os-net-config-dcb entry
points are updated as well.

(cherry picked from commit 7dc3da24f13da88de84c6d60683f5e69b32146c8)
Signed-off-by: Abhiram R N <abhiramrn@gmail.com>