udevadm: handle udevadm wait on RHEL8.4
In RHEL8.4 the `udevadm wait` is not supported. The command
is run so as to wait for the sysfs to be ready for configuring
sriov_numvfs. So when `udevadm wait fails`, an attempt is made
using `udevadm settle` to wait for the udev events.

(cherry picked from commit f63c6f07804e2feef4cb3dd6c6284e3bacf60ee3)
Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>