sriov: Apply the PF/VF config only when there is a change
The desired and the current states of the devices are compared
for PFs and VFs. The desired state is applied only when there is
a change in configuration. A ticket is opened in nmstate to perform
the comparison of the states and apply it when needed.
JIRA: https://issues.redhat.com/browse/RHEL-67120

(cherry picked from commit 5bc677adf1e695bf2566adb55bb9bc3a5cbd76cc)
Signed-off-by: Abhiram R N <abhiramrn@gmail.com>