logs: make logging precise and readable
The logging in nmstate provider is printing the data structures
in json format which is not easy to read and decipher. These data
structures are printed in yaml format for the ease of reading.
Also some of the logs are repeated at several levels leading to
duplication of the logging. These issues are addressed here.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>