utils: simplify get_pci_address() implementation
Presently the PCI address is read using ethtool command, when the
device is bound with standard netdev driver. It is fetched from dpdk
map when its bound with vfio-pci. The caller is expected to identify
the device type and choose the appropriate function.
The get_pci_address(interface_name) will now handle fetching the PCI
address of the interface irrespective of the device types. In addition
to that, now the PCI address of the SR-IOV VFs also could be fetched
by passing the interface name in the format f"{pf_name}:{vfid}".
With the usage of sysfs, the name of sriov vf is not required to fetch
the pci address. This will be used in nic partitioning use cases when
sriov_drivers_autoprobe is disabled.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>