sriov: skip PF configuration when its attached to guest
In scenarios where the PF is attached to the guest, the PF device
will not be visible in the hosts net subsystem. And if os-net-config
is rerun in that scenario, the NIC mapping would fail. Checks shall
be added so that the situation where the PF is attached to the guest
is detected and handled.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>