schema: route schema shall support nexthop and next_hop
The current route schema requires `next_hop` when `ip_netmask` is
configured and `nexthop` when `destination` is configured. But since the
schema validation is ignored until recently and the object parser parses
both `nexthop` and `next_hop` alike, the schema validation failure for
the same did not fail os-net-config. Now the schema for `route` object
is modified to support both `nexthop` and `next_hop` and hence the
schema validation will not fail on existing deployments for the same.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>