bond: ignore order for dpdkbond
In ifcfg provider, the utils.diff() compares the content of the new ifcfg file
and existing ifcfg file.
For OvS devices, there are fields like BOND_IFACES, OVS_EXTRA where the order of the entry
doesn’t matter. The diff shall ignore the order of the entries for such fields.

Closes: https://issues.redhat.com/browse/OSPRH-13695
Signed-off-by: Vijayalakshmi Candappa <vcandapp@redhat.com>