cleanup: stop blindly removing nmconnection files
Stops removing NetworkManager connections from the ifcfg
provider. The logic in the NetworkManager cleanup is too
indiscriminate and can cause removal of connections that
don't need to be removed, should not be removed, and/or
can cause spurious errors. A follow-up patch will
ensure there are no conflicts in a safe way that allows
for peaceful co-existence of network-scripts and nmstate.

Suggested-by: Karthik Sundaravel <ksundara@redhat.com>
Signed-off-by: Dan Sneddon <dsneddon@redhat.com>