nmstate: purge nmstate provider
Remove the network configurations made by nmstate provider.
All the network devices in config.yaml would be destroyed and the
devices would be brought back to clean state. For DPDK ports the interfaces
would be detached and unbound from vfio-pci driver. In case of SR-IOV devices
the numvfs would be reset and the VFs will be removed. The dpdk, sriov
map files would be backed up.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>