nmstate: avoid cleanup of ovs interfaces and sriov pfs
The cleanup actions cannot be allowed on ovs interfaces, bridges and
SR-IOV devices as removing these network interfaces will cause
significant down time to the networks and some of these interfaces
cannot be removed either. If there is a real need to remove these
interfaces it would be recommended to re-configure the system again.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>