min_config: add minimum config support
Adding support for running mnimum configuration. The minimum_config
shall use the same template format as `network_config` section.
The minimum_config when present in `config.yaml` shall be applied before
applying the `network_config` section. This is to ensure that the bare
minimum configuration is applied before running the `network_config`.
The minimum config support is required for migration as well as applying
a bare minimum configuration.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>