ifcfg: add helpers to get pci address of DPDK ports / bonds
The remove_config feature needs the helper methods to fetch the pci address
of the DPDK ports and the DPDK bonds. The pci address fetched could be later
used in detaching it from the dpdk ports and unbind the same from vfio-pci
driver. Since the remove_config feature could remove a dpdk interface which
is not previously configured by os-net-config, the usage of dpdk map is
avoided. The ifcfg files for the respective ports or bonds are instead used
in reading the PCI addresses.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>