ifcfg: set sysctl to set keep_addr_on_down
Write a sysctl.d config file in /usr/lib/sysctl.d when
IPv6 addresses are used to set keep_addr_on_down as a
default and for all existing interfaces. This behavior
is default for IPv4 addresses and needs to be used for
IPv6 addresses on systems that are routing IPv6.

Signed-off-by: Dan Sneddon <dsneddon@redhat.com>